+++ web/src/scripts/pages/GamePage.ts
@@ -1,6 +1,6 @@
 import { createHeader } from "../components/Header";
 import { createFooter } from "../components/Footer";
-import { createGameSection } from "../components/GameSection";
+import { createGameSection } from "../components/GameSection3D";
 
 export const createGamePage = async (): Promise<HTMLElement> => {
     const fragment = document.createDocumentFragment();
